[BACKLOG-38674] Integration Tests are outdated and some do not compile.
excel*
