+++ plugins/excel/core/src/it/java/org/pentaho/di/trans/steps/excelinput/OdsWorkBookIT.java
@@ -22,18 +22,17 @@
 
 package org.pentaho.di.trans.steps.excelinput;
 
-import java.util.Date;
-
 import junit.framework.TestCase;
-
 import org.pentaho.di.core.spreadsheet.KCell;
 import org.pentaho.di.core.spreadsheet.KCellType;
 import org.pentaho.di.core.spreadsheet.KSheet;
 import org.pentaho.di.core.spreadsheet.KWorkbook;
 
-public class JxlWorkBookIT extends TestCase {
+import java.util.Date;
+
+public class OdsWorkBookIT extends TestCase {
   public void testRead() throws Exception {
-    KWorkbook workbook = WorkbookFactory.getWorkbook( SpreadSheetType.JXL, "src/it/resources/sample-file.xls", null );
+    KWorkbook workbook = WorkbookFactory.getWorkbook( SpreadSheetType.ODS, "src/it/resources/sample-file.ods", null );
     int numberOfSheets = workbook.getNumberOfSheets();
     assertEquals( 3, numberOfSheets );
     KSheet sheet1 = workbook.getSheet( 0 );
@@ -50,7 +49,7 @@ public class JxlWorkBookIT extends TestCase {
     assertEquals( new Date( 1283817600000L ), row[2].getValue() );
     assertEquals( KCellType.NUMBER, row[3].getType() );
     assertEquals( Double.valueOf( "75" ), row[3].getValue() );
-    assertEquals( KCellType.BOOLEAN, row[4].getType() );
+    assertEquals( KCellType.BOOLEAN_FORMULA, row[4].getType() ); // Always like that in ODS
     assertEquals( Boolean.valueOf( true ), row[4].getValue() );
     assertEquals( KCellType.NUMBER_FORMULA, row[5].getType() );
     assertEquals( Double.valueOf( "75" ), row[5].getValue() );
@@ -62,7 +61,7 @@ public class JxlWorkBookIT extends TestCase {
     assertEquals( new Date( 1283904000000L ), row[2].getValue() );
     assertEquals( KCellType.NUMBER, row[3].getType() );
     assertEquals( Double.valueOf( "42" ), row[3].getValue() );
-    assertEquals( KCellType.BOOLEAN, row[4].getType() );
+    assertEquals( KCellType.BOOLEAN_FORMULA, row[4].getType() ); // Always like that in ODS
     assertEquals( Boolean.valueOf( false ), row[4].getValue() );
     assertEquals( KCellType.NUMBER_FORMULA, row[5].getType() );
     assertEquals( Double.valueOf( "117" ), row[5].getValue() );
@@ -74,17 +73,10 @@ public class JxlWorkBookIT extends TestCase {
     assertEquals( new Date( 1283990400000L ), row[2].getValue() );
     assertEquals( KCellType.NUMBER, row[3].getType() );
     assertEquals( Double.valueOf( "93" ), row[3].getValue() );
-    assertEquals( KCellType.BOOLEAN, row[4].getType() );
+    assertEquals( KCellType.BOOLEAN_FORMULA, row[4].getType() ); // Always like that in ODS
     assertEquals( Boolean.valueOf( true ), row[4].getValue() );
     assertEquals( KCellType.NUMBER_FORMULA, row[5].getType() );
     assertEquals( Double.valueOf( "210" ), row[5].getValue() );
-
-    try {
-      sheet1.getRow( 5 );
-      throw new Exception( "No out of bounds exception thrown when expected" );
-    } catch ( ArrayIndexOutOfBoundsException e ) {
-      // OK!
-    }
 
     workbook.close();
   }
