+++ plugins/excel/core/src/it/resources/sample-file.xlsx
